Move spec files to acceptance.
