+++ bigtable/Rakefile
@@ -12,15 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-source "https://rubygems.org"
+require "rake/testtask"
 
-git_source(:github) { |repo_name| "https://github.com/#{repo_name}" }
-
-gem "google-cloud-bigtable", "~> 1.0", ">= 1.0.2"
-
-group :test do
-  gem "minitest", "~> 5.13"
-  gem "minitest-focus", "~> 1.1"
-  gem "minitest-junit"
-  gem "rake"
+Rake::TestTask.new "test" do |t|
+  t.test_files = FileList["**/*_test.rb"]
+  t.options = "--junit --junit-filename=sponge_log.xml"
+  t.warning = false
 end
