+++ bigtable/acceptance/filter_samples_test.rb
