+++ bigtable/acceptance/quickstart_test.rb
