+++ bigtable/acceptance/read_samples_test.rb
