+++ bigtable/acceptance/write_samples_test.rb
